Store fingerprint as raw byte array

+++ ssl/crypto_misc.h
@@ -76,7 +76,7 @@ struct _x509_ctx
     uint8_t sig_type;
     RSA_CTX *rsa_ctx;
     bigint *digest;
-    bigint *fingerprint;
+    uint8_t *fingerprint;
     struct _x509_ctx *next;
 };
